# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plataforma de Cursos",
        "description": "API para gestionar cursos, instructores, lecciones, comentarios, calificaciones y favoritos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/users/{user}/favorite-courses/{course}": {
            "post": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Marca curso como favorito",
                "operationId": "favoriteCourse",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Curso marcado como favorito"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Desmarca curso como favorito",
                "operationId": "unfavoriteCourse",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Curso removido de favoritos"
                    }
                }
            }
        },
        "/api/courses/{course}/comments": {
            "post": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Comenta un curso",
                "operationId": "commentCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "body"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Excelente curso de Laravel."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comentario creado"
                    }
                }
            }
        },
        "/api/instructors/{instructor}/comments": {
            "post": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Comenta un instructor",
                "operationId": "commentInstructor",
                "parameters": [
                    {
                        "name": "instructor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "body"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Muy buen instructor."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comentario creado"
                    }
                }
            }
        },
        "/api/courses/{course}/ratings": {
            "post": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Califica un curso",
                "operationId": "rateCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "score"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "score": {
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "review": {
                                        "type": "string",
                                        "example": "Muy completo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Calificacion registrada"
                    }
                }
            }
        },
        "/api/instructors/{instructor}/ratings": {
            "post": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Califica un instructor",
                "operationId": "rateInstructor",
                "parameters": [
                    {
                        "name": "instructor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "score"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "score": {
                                        "type": "integer",
                                        "example": 4,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "review": {
                                        "type": "string",
                                        "example": "Buen nivel de explicacion"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Calificacion registrada"
                    }
                }
            }
        },
...
    },
...
}